Refactor into a base renderer with view renderer subclass
