+++ pakyow-presenter/lib/pakyow/presenter/presenter.rb
@@ -11,7 +11,6 @@ require "pakyow/support/pipelined"
 require "pakyow/support/pipelined/haltable"
 
 require "pakyow/presenter/errors"
-require "pakyow/presenter/renderer"
 
 require "pakyow/presenter/presenter/behavior/endpoints"
