+++ pakyow-presenter/lib/pakyow/presenter/rendering/actions/cleanup_prototype_nodes.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Pakyow
+  module Presenter
+    module Actions
+      # @api private
+      class CleanupPrototypeNodes
+        def initialize(_options)
+        end
+
+        def call(renderer)
+          unless renderer.rendering_prototype?
+            renderer.presenter.view.object.find_significant_nodes(:prototype).each(&:remove)
+          end
+        end
+      end
+    end
+  end
+end
